% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_list}
\alias{append_list}
\title{Append two lists}
\usage{
append_list(x, y)
}
\arguments{
\item{x}{List base}

\item{y}{Second list}
}
\value{
x
}
\description{
Get commom parameters in two list and generate one append list
}
