% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_summary.R
\name{chain_summary}
\alias{chain_summary}
\title{Summary measures for a(n) (un)restricted model}
\usage{
chain_summary(obj, level)
}
\arguments{
\item{obj}{A matrix containing parameter's chains}

\item{level}{Credibility level}
}
\value{
Measures
}
\description{
Summary measures for a(n) (un)restricted model
}
