% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_glmm.R
\name{control_family_dft}
\alias{control_family_dft}
\title{Default values to control GLMM families}
\usage{
control_family_dft(family)
}
\arguments{
\item{family}{"gaussian", "poisson", "binomial".}
}
\description{
Default values to control GLMM families
}
