% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{cum_hexpsurv}
\alias{cum_hexpsurv}
\title{cum_hexpsurv}
\usage{
cum_hexpsurv(t, rate)
}
\arguments{
\item{t}{time vector.}

\item{rate}{rate vector.}
}
\description{
Exponential cumulative hazards
}
\keyword{internal}
