% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{cum_hpwesurv}
\alias{cum_hpwesurv}
\title{cum_hpwesurv}
\usage{
cum_hpwesurv(t, tgrid, rates)
}
\arguments{
\item{t}{Time's vector}

\item{tgrid}{Time's grid}

\item{rates}{Exponential rates}
}
\description{
Piecewise Exponential cumulative hazzards
}
