% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{cum_hweibullsurv}
\alias{cum_hweibullsurv}
\title{cum_hweibullsurv}
\usage{
cum_hweibullsurv(t, alpha, lambda)
}
\arguments{
\item{t}{time vector.}

\item{alpha}{alpha parameter.}

\item{lambda}{lambda parameter.}
}
\description{
Weibull cumulative hazards
}
\keyword{internal}
