% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(object)
}
\arguments{
\item{object}{a object from ?rsglmm, ?rscm or ?rsfm}
}
\value{
DIC
}
\description{
Get the Deviance Information Criterion (DIC) from a model
}
