% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sc.R
\name{gamma_prior}
\alias{gamma_prior}
\title{gamma_prior}
\usage{
gamma_prior(mean, var, a = NULL, b = NULL)
}
\arguments{
\item{mean}{Mean of a desired gamma distribution (restore a and b)}

\item{var}{Variance of a desired gamma distribution (restore a and b)}

\item{a}{Shape of a desired gamma distribution (restore mean and variance)}

\item{b}{Scale of a desired gamma distribution (restore mean and variance)}
}
\description{
Given mean and variance return a and b
}
