% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sc.R
\name{get_trans_samp}
\alias{get_trans_samp}
\title{get_trans_samp}
\usage{
get_trans_samp(marg, fun, n = 1000, trunc = FALSE, method = "linear")
}
\description{
Generating a sample from the transformed marginal
}
