% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{hpwesurv}
\alias{hpwesurv}
\title{hpwesurv}
\usage{
hpwesurv(t, rates, tgrid)
}
\arguments{
\item{t}{Time's vector}

\item{rates}{Exponential rates}

\item{tgrid}{Time's grid}
}
\description{
Piecewise Eponential hazzards
}
