% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neigh_RJ.R
\docType{data}
\name{neigh_RJ}
\alias{neigh_RJ}
\title{Rio de Janeiro's shapefile}
\format{
An object of class \code{"SpatialPolygonsDataFrame"}.
}
\usage{
data(neigh_RJ)
}
\description{
Rio de Janeiro's cities shape file
}
\keyword{datasets}
