% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_mat}
\alias{proj_mat}
\title{Projection matrix}
\usage{
proj_mat(X, groups = NULL, method = "rhz")
}
\arguments{
\item{X}{covariate matrix}

\item{groups}{subject id}

\item{method}{specific projection method, for now just "rhz" is available.}
}
\description{
Calculate the projection matrix under several approaches
}
\keyword{internal}
