% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcar}
\alias{rcar}
\title{Generate data from CAR model}
\usage{
rcar(W, sig, rho = 0.9999)
}
\arguments{
\item{W}{Adjcency matrix}

\item{sig}{Standard deviation}

\item{rho}{Dependence parameter}
}
\value{
x
}
\description{
Generate data from CAR model
}
