% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ricar}
\alias{ricar}
\title{Generate data from ICAR model}
\usage{
ricar(W, sig = 1)
}
\arguments{
\item{W}{Adjcency matrix}

\item{sig}{Standard deviation}
}
\value{
x
}
\description{
Generate data from ICAR model
}
