% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscm.R
\name{rscm}
\alias{rscm}
\title{Restricted Shared Component Model}
\usage{
rscm(
  data,
  Y1,
  Y2,
  X1,
  X2,
  E1 = NULL,
  E2 = NULL,
  W,
  neigh,
  area,
  proj = "none",
  nsamp = 1000,
  ...
)
}
\arguments{
\item{data}{data.frame}
}
\value{
Restricted model
}
\description{
Fit a Restricted Shared Component Model
}
