% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm.R
\name{rsfm}
\alias{rsfm}
\title{Restricted Spatial Frailty Model}
\usage{
rsfm(data, time, time2 = NULL, status, covariates, intercept = TRUE, area = NULL,
            model = NULL, W = NULL,
            family, cure = FALSE, proj = "none", fast = TRUE, nsamp = 1000,
            approach = "inla", ...)
}
\arguments{
\item{data}{data.frame}

\item{time}{For right censored data, this is the follow up time. For interval data, this is the starting time for the interval.}

\item{time2}{Ending time for the interval censured data.}

\item{status}{The status indicator, 1 = observed event, 0 = right censored event, 2 = left censored event, 3 = interval censored event.}

\item{covariates}{Vector of covariates names}

\item{intercept}{TRUE or FALSE}

\item{area}{Column representing the region of each individual}

\item{model}{Spatial model adopted}

\item{W}{Adjacency matrix}

\item{family}{'exponential', 'weibull', 'weibullcure', 'loglogistic', 'gamma', 'lognormal' or 'pwe'}

\item{cure}{Cure fraction? TRUE or FALSE}

\item{proj}{'none', 'rhz', 'hh' or 'spock'}

\item{fast}{To use the reduction operator}

\item{nsamp}{Sample size to use the projection approach}

\item{approach}{'inla' or 'bugs'}

\item{...}{Other parameters used in ?inla or ?bugs}
}
\value{
Restricted model
}
\description{
Fit a Restricted Spatial Frailty model
}
