% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm_bugs.R
\name{rsfm_bugs}
\alias{rsfm_bugs}
\title{Restricted Spatial Frailty Model in BUGS}
\usage{
rsfm_bugs(
  model,
  data,
  inits,
  parameters,
  covariates,
  area,
  proj = "none",
  fast = TRUE,
  nsamp = 1000,
  burnin = 5000,
  thin = 10,
  ...
)
}
\arguments{
\item{model}{character with the text model}

\item{data}{dataset}

\item{inits}{initial values for the chain}

\item{parameters}{Vector of parameters to restore}

\item{covariates}{Vector of covariates names}

\item{proj}{'none', 'rhz', 'hh' or 'spock'}

\item{fast}{To use the reduction operator}

\item{nsamp}{Sample size to use the projection approach}

\item{burnin}{Burnin period}

\item{thin}{Lag parameter}

\item{...}{Other parameters used in ?inla}
}
\value{
INLA object with corrected parameters
}
\description{
Fit a Restricted Spatial Frailty model using INLA
}
