% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm_inla.R
\name{rsfm_inla}
\alias{rsfm_inla}
\title{Restricted Spatial Frailty Model in INLA}
\usage{
rsfm_inla(data, formula, family, W = NULL,
                 proj = "none", nsamp = 1000, fast = TRUE,
                 ...)
}
\arguments{
\item{data}{an data frame or list containing the variables in the model.}

\item{formula}{a inla formula like inla.surv(time, event) ~ 1 + z + f(ind, model="iid") + f(ind2, weights, model="ar1"). This is much like the formula for a glm except that smooth or spatial terms can be added to the right hand side of the formula. See f for full details and the web site www.r-inla.org for several worked out examples. Each smooth or spatial term specified through f should correspond to separate column of the data frame data. The outcome is the output of the function inla.surv.}

\item{family}{"exponential", "weibull", "weibullcure", "loglogistic", "gamma", "lognormal" or "pwe".}

\item{W}{adjacency matrix.}

\item{proj}{"none" or "rhz".}

\item{nsamp}{number of samples. Default = 1000.}

\item{fast}{TRUE to use the reduction operator.}

\item{...}{other parameters used in ?INLA::inla.}
}
\value{
\item{$unrestricted}{A list containing
                               \itemize{
                                  \item $sample: a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed: summary measures for the coefficients
                                  \item $summary_hyperpar: summary measures for hyperparameters
                                  \item $summary_random: summary measures for random quantities
                                }
                             }
\item{$restricted}{A list containing
                               \itemize{
                                  \item $sample: a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed: summary measures for the coefficients
                                  \item $summary_hyperpar: summary measures for hyperparameters
                                  \item $summary_random: summary measures for random quantities
                                }
                             }

\item{$out}{INLA output}
\item{$time}{time elapsed for fitting the model}
}
\description{
Fit a Restricted Spatial Frailty model using INLA
}
