% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsfm_inla_old}
\alias{rsfm_inla_old}
\title{Restricted Spatial Frailty Model in INLA}
\usage{
rsfm_inla_old(
  f,
  data,
  family,
  cure = FALSE,
  proj = "none",
  fast = TRUE,
  nsamp = 1000,
  ...
)
}
\arguments{
\item{f}{INLA formula ?inla.surv}

\item{data}{data.frame}

\item{family}{'exponential', 'weibull', 'weibullcure', 'loglogistic', 'gamma', 'lognormal' or 'pwe'}

\item{cure}{Desired cure fraction}

\item{proj}{'none', 'rhz', 'hh' or 'spock'}

\item{fast}{To use the reduction operator}

\item{nsamp}{Sample size to use the projection approach}

\item{...}{Other parameters used in ?inla}
}
\value{
INLA object with corrected parameters
}
\description{
Fit a Restricted Spatial Frailty model using INLA
}
