% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsglmm.R
\name{rsglmm}
\alias{rsglmm}
\title{Restricted Spatial Generalized Linear Mixed model}
\usage{
rsglmm(data, Y, covariates, intercept = TRUE, area = NULL,
              neigh = NULL,
              family, proj = "none", nsamp = 1000,
              approach = "inla", ...)
}
\value{
Restricted model
}
\description{
Fit a Restricted Spatial Generalized Linear Mixed model
}
