% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsglmm_mcmc.R
\name{rsglmm_mcmc}
\alias{rsglmm_mcmc}
\title{Restricted Spatial Generalized Linear Mixed model}
\usage{
rsglmm_mcmc(
  data,
  formula,
  family,
  W,
  area,
  proj,
  nsamp = 1000,
  burnin = 5000,
  lag = 1,
  attractive = round(0.5 * (nrow(W)/2)),
  ...
)
}
\arguments{
\item{data}{an data frame or list containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{allowed families are: 'gaussian', 'poisson' and 'binomial'.}

\item{W}{adjacency matrix.}

\item{area}{areal variable name in \code{data}.}

\item{proj}{'hh'}

\item{nsamp}{number of desired. samples Default = 1000.}

\item{burnin}{burnin size.}

\item{lag}{lag parameter.}

\item{attractive}{the number of attractive Moran eigenvectors to use. See ?ngspatial::sparse.sglmm for more information.}

\item{...}{other parameters used in ?ngspatial::sparse.sglmm}
}
\value{
\item{$unrestricted}{A list containing
                               \itemize{
                                  \item $sample a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed summary measures for the coefficients
                                  \item $summary_hyperpar summary measures for hyperparameters
                                  \item $summary_random summary measures for random quantities
                                }
                             }
\item{$restricted}{A list containing
                               \itemize{
                                  \item $sample a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed summary measures for the coefficients
                                  \item $summary_hyperpar summary measures for hyperparameters
                                  \item $summary_random summary measures for random quantities
                                }
                             }

\item{$out}{ngspatial output}
\item{$time}{time elapsed for fitting the model}
}
\description{
Fit a Restricted Spatial Generalized Linear Mixed model using ngspatial
}
