% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rsurv}
\alias{rsurv}
\title{rsurv}
\usage{
rsurv(
  n_id,
  coefs = c(0.1, 0.4, -0.3),
  cens = 0,
  cens_type = "interval",
  hazard = "weibull",
  hazard_params = hazard_dft(),
  spatial = "ICAR",
  neigh = NULL,
  W = NULL,
  tau = 1,
  confounding = "none",
  proj = "none",
  sd_x = 0,
  X = NULL,
  scale = TRUE
)
}
\arguments{
\item{n_id}{vector with the number of individuals in each region to be generated.}

\item{coefs}{vector of coefficients.}

\item{cens}{proportion of censoring.}

\item{cens_type}{censoring scheme: "none", "left", "right" or "interval".}

\item{hazard}{hazard model: "exponential", "weibull" or "pwe" for piecewise exponential.}

\item{hazard_params}{named list with parameters for the hazard model: hazard_dft().}

\item{spatial}{spatial model: "none" for the conventional Cox model,
"gamma" for an independent gamma frailty,
"lognormal" for an independent lognormal frailty,
"ICAR" or "BYM" for spatial structured models.}

\item{neigh}{neighborhood structure. A \code{SpatialPolygonsDataFrame} object.}

\item{W}{adjacency matrix.}

\item{tau}{precision for ICAR and BYM models.}

\item{confounding}{"none", "linear", "quadratic" or "cubic".}

\item{proj}{"none", "rhz", "hh" or "spock".}

\item{sd_x}{standard deviation to generating confounding.}

\item{X}{matrix of covariates. Default = NULL.}

\item{scale}{scale X. TRUE or FALSE.}
}
\description{
Generating data with spatial confounding
}
