% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rsurv_interval}
\alias{rsurv_interval}
\title{rsurv_interval}
\usage{
rsurv_interval(times, cens)
}
\arguments{
\item{times}{observed time vector.}

\item{cens}{censored proportion.}
}
\description{
Generating data with interval censoring
}
\keyword{internal}
