% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_marginal}
\alias{select_marginal}
\title{select_marginal}
\usage{
select_marginal(samp, ids)
}
\arguments{
\item{samp}{a sample from ?inla.posterior.sample}

\item{ids}{ids to restore}
}
\description{
Select the desired marginals on a INLA models
}
\keyword{internal}
