% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spock.R
\name{spock}
\alias{spock}
\title{SPOCK nb}
\usage{
spock(X, map)
}
\arguments{
\item{X}{covariate matrix.}

\item{map}{object of class \code{nb}.}
}
\value{
map_new object of class \code{nb}
}
\description{
SPOCK approach to transform the neighborhood structure
}
