% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{surv}
\alias{surv}
\title{surv}
\usage{
surv(time, time2 = NULL, event)
}
\arguments{
\item{time}{Time until event (censure) or lower limit for interval censoring}

\item{time2}{Upper limit for interval censoring}

\item{event}{The event indicator, 1 = observed event, 0 = right censored event, 2 = left censored event, 3 = interval censored event.}
}
\description{
Auxiliar function for survival models
}
