% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVIF}
\alias{SVIF}
\title{Spatial Variance Inflation Factor}
\usage{
SVIF(base_model, model)
}
\arguments{
\item{base_model}{Model to use as comparison}

\item{model}{Model to compare}
}
\value{
res data.frame with VIF for fixed parameters
}
\description{
Calculate the VIF
}
