% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVRF}
\alias{SVRF}
\title{Spatial Variance Retraction Factor}
\usage{
SVRF(base_model, model)
}
\arguments{
\item{base_model}{Model to use as comparison}

\item{model}{Model to compare}
}
\value{
res data.frame with VRF for fixed parameters
}
\description{
Calculate the VRF
}
