% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_inla_formula}
\alias{update_inla_formula}
\title{Updating INLA formula}
\usage{
update_inla_formula(formula)
}
\arguments{
\item{formula}{a formula to be updated to INLA format}
}
\description{
Updating INLA formula
}
\keyword{internal}
