% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WAIC}
\alias{WAIC}
\title{Watanabe–Akaike information criterion}
\usage{
WAIC(object)
}
\arguments{
\item{object}{a object from ?rsglmm, ?rscm or ?rsfm}
}
\value{
WAIC
}
\description{
Get the Watanabe–Akaike information criterion (WAIC) from a model
}
\references{
Watanabe, S. (2010). \emph{"Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in Singular Learning Theory"}. Journal of Machine Learning Research, 11, 3571-3594.
}
