% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_summary.R
\name{chain_summary}
\alias{chain_summary}
\title{Summary measures for a(n) (un)restricted model}
\usage{
chain_summary(obj, level = 0.95)
}
\arguments{
\item{obj}{An restricted model chain output}
}
\value{
Measures
}
\description{
Summary measures for a(n) (un)restricted model
}
