% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_summary.R
\name{chain_summary}
\alias{chain_summary}
\title{Summary measures for a sample}
\usage{
chain_summary(sample, level)
}
\arguments{
\item{sample}{a matrix containing parameter's chains.}

\item{level}{credibility interval level.}
}
\value{
a matrix with \code{mean}, \code{median}, \code{sd}, \code{lower} and \code{upper} measures
}
\description{
Summary measures for a sample
}
