% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{cum_hpwesurv}
\alias{cum_hpwesurv}
\title{cum_hpwesurv}
\usage{
cum_hpwesurv(t, tgrid, rates)
}
\arguments{
\item{t}{time vector.}

\item{tgrid}{time grid.}

\item{rates}{exponential rates.}
}
\description{
Piecewise Exponential cumulative hazards
}
\keyword{internal}
