% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{cum_hweibullsurv}
\alias{cum_hweibullsurv}
\title{cum_hweibullsurv}
\usage{
cum_hweibullsurv(t, alpha, lambda)
}
\arguments{
\item{t}{Time's vector}

\item{alpha}{Alpha parameter}

\item{lambda}{Lambda parameter}
}
\description{
Weibull cumulative hazzards
}
