% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sc.R
\name{gamma_prior}
\alias{gamma_prior}
\title{gamma_prior}
\usage{
gamma_prior(mean, var, a = NULL, b = NULL)
}
\arguments{
\item{mean}{mean of a desired gamma distribution (restore a and b).}

\item{var}{variance of a desired gamma distribution (restore a and b).}

\item{a}{shape of a desired gamma distribution (restore mean and variance).}

\item{b}{scale of a desired gamma distribution (restore mean and variance).}
}
\description{
Given mean and variance return a and b
}
\keyword{internal}
