% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{hpwesurv}
\alias{hpwesurv}
\title{hpwesurv}
\usage{
hpwesurv(t, rates, tgrid)
}
\arguments{
\item{t}{time vector.}

\item{rates}{exponential rates.}

\item{tgrid}{time grid.}
}
\description{
Piecewise Eponential hazards
}
\keyword{internal}
