% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{hweibullsurv}
\alias{hweibullsurv}
\title{hweibullsurv}
\usage{
hweibullsurv(t, alpha, lambda)
}
\arguments{
\item{t}{Time's vector}

\item{alpha}{Alpha parameter}

\item{lambda}{Lambda parameter}
}
\description{
Weibull hazzards
}
