% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{hweibullsurv}
\alias{hweibullsurv}
\title{hweibullsurv}
\usage{
hweibullsurv(t, alpha, lambda)
}
\arguments{
\item{t}{time vector.}

\item{alpha}{alpha parameter.}

\item{lambda}{lambda parameter.}
}
\description{
Weibull hazards
}
\keyword{internal}
