% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meang}
\alias{meang}
\title{meang}
\usage{
meang(x, g, weighted = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{g}{Group indexes}

\item{weighted}{TRUE for weighted mean}
}
\description{
Mean by groups
}
