% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_mat}
\alias{proj_mat}
\title{Projection matrix}
\usage{
proj_mat(X, groups = NULL, method = "rhz")
}
\arguments{
\item{X}{Covariate matrix}

\item{groups}{Ids for the subjects}

\item{method}{Projection's method}
}
\value{
Px Projection matrix
Px_ort (I - Px)
}
\description{
Calculate the projection matrix under several approaches
}
