% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rexpsurv}
\alias{rexpsurv}
\title{rexpsurv}
\usage{
rexpsurv(N, rate = rep(1, N))
}
\arguments{
\item{N}{number of observations.}

\item{rate}{rate vector.}
}
\description{
Generate data from Exponential hazard model
}
\keyword{internal}
