% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_glmm.R
\name{rglmm}
\alias{rglmm}
\title{rglmm}
\usage{
rglmm(
  beta = c(0.1, -0.1),
  tau = 1,
  family = "gaussian",
  confounding = "none",
  neigh,
  scale = TRUE,
  control_family = NULL
)
}
\arguments{
\item{beta}{Coefficients parameters}

\item{tau}{Precision for ICAR}

\item{family}{"gaussian", "binomial" or "poisson"}

\item{confounding}{'none', 'linear', 'quadratic' or 'cubic'}

\item{neigh}{Neighborhood structure. A \code{SpatialPolygonsDataFrame} object}

\item{scale}{Scale covariates? TRUE or FALSE}

\item{control_family}{A list with, at least, a \code{invlink}}
}
\description{
Generating data from Generalized Linear Mixed models
}
