% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rpwesurv}
\alias{rpwesurv}
\title{rpwesurv}
\usage{
rpwesurv(N, effects = rep(1, N), rates, tgrid)
}
\arguments{
\item{N}{number of observations.}

\item{effects}{effects.}

\item{rates}{exponential rates.}

\item{tgrid}{time's grid.}
}
\description{
Generate data from Piecewise Exponential hazzard model
}
