% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm.R
\name{rsfm}
\alias{rsfm}
\title{Restricted Spatial Frailty Model}
\usage{
rsfm(data, formula, area = NULL,
            model = NULL, neigh = NULL,
            family, proj = "none", fast = TRUE, nsamp = 1000,
            approach = "inla", ...)
}
\arguments{
\item{data}{data.frame containing, at least, \code{time}, \code{event}, \code{covariates}, \code{area} list}

\item{formula}{A formula in the format surv(time, time2, event) ~ X1 + X2}

\item{area}{Column of data specifying the region of each individual}

\item{model}{Spatial model adopted: "besag" or  "restricted_besag"}

\item{neigh}{Neighborhood structure. A \code{SpatialPolygonsDataFrame} object}

\item{family}{'exponential', 'weibull', 'weibullcure', 'loglogistic', 'gamma', 'lognormal' or 'pwe'}

\item{proj}{'none', 'rhz' or 'spock'}

\item{fast}{To use the reduction operator}

\item{nsamp}{Sample size to use the projection approach}

\item{approach}{'inla' or 'bugs'}

\item{...}{Other parameters used in ?inla or ?bugs}
}
\value{
\item{$unrestricted}{A list containing $sample, $summary_fixed, $summary_hyperpar, $summary_random, $out and $time}
\item{$restricted}{A list containing $sample, $summary_fixed, $summary_hyperpar, $summary_random, $out and $time}
}
\description{
Fit a Restricted Spatial Frailty model
}
\examples{
set.seed(1)

##-- Spatial structure
data("neigh_RJ")

##-- Individuals and regions
n_reg <- length(neigh_RJ)
n_id <- sample(x = 3:5, size = n_reg, replace = TRUE)
coefs <- c(0.3, -0.3)
tau <- 0.75 # Scale of spatial effect

##-- Data
data <- rsurv(n_id = n_id,
              coefs = coefs, cens = 0.5, scale = FALSE,
              cens_type = "right", hazard = "weibull",
              hazard_params = hazard_params <- list(weibull = list(alpha = 1.2, variant = 0)),
              spatial = "ICAR",
              neigh = neigh_RJ, tau = tau, confounding = "linear", proj = "none")

##-- Models
weibull_inla <- rsfm(data = data,
                     formula = surv(time = L, event = status) ~ X1 + X2,
                     model = "none", family = "weibull",
                     proj = "rhz", nsamp = 1000, approach = "inla")

rsfm_inla <- rsfm(data = data, area = "reg",
                  formula = surv(time = L, event = status) ~ X1 + X2,
                  model = "restricted_besag", neigh = neigh_RJ, family = "weibull",
                  proj = "rhz", nsamp = 1000, approach = "inla")

weibull_inla$unrestricted$summary_fixed
rsfm_inla$unrestricted$summary_fixed
rsfm_inla$restricted$summary_fixed

}
