% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm.R
\name{rsfm}
\alias{rsfm}
\title{Restricted Spatial Frailty Model}
\usage{
rsfm(data, formula, family, area = NULL,
            model = NULL, neigh = NULL,
            proj = "none", nsamp = 1000,
            fast = TRUE, approach = "inla", priors,
            ...)
}
\arguments{
\item{data}{a data frame or list containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{"exponential", "weibull", "weibullcure", "loglogistic", "gamma", "lognormal" or "pwe".}

\item{area}{areal variable name in \code{data}.}

\item{model}{spatial model adopted. Examples: "besag", "besag2" or "r_besag". See INLA::inla.list.models() for other models.}

\item{neigh}{neighborhood structure. A \code{SpatialPolygonsDataFrame} object.}

\item{proj}{"none" or "rhz".}

\item{nsamp}{number of samples. Default = 1000.}

\item{fast}{TRUE to use the reduction operator.}

\item{approach}{"inla" or "mcmc". "mcmc" has less model options.}

\item{priors}{a list containing:
\itemize{
   \item prior_prec: a vector of size two, representing shape and scale for the gamma distribution applied for model precision
}}

\item{...}{other parameters used in ?INLA::inla or ?R2OpenBUGS::bugs}
}
\value{
\item{$unrestricted}{A list containing
                               \itemize{
                                  \item $sample: a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed: summary measures for the coefficients
                                  \item $summary_hyperpar: summary measures for hyperparameters
                                  \item $summary_random: summary measures for random quantities
                                }
                             }
\item{$restricted}{A list containing
                               \itemize{
                                  \item $sample: a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed: summary measures for the coefficients
                                  \item $summary_hyperpar: summary measures for hyperparameters
                                  \item $summary_random: summary measures for random quantities
                                }
                             }

\item{$out}{INLA or BUGS output}
\item{$time}{time elapsed for fitting the model}
}
\description{
Fit a Restricted Spatial Frailty model
}
\examples{
set.seed(123456)

##-- Spatial structure
data("neigh_RJ")

##-- Individuals and regions
n_reg <- length(neigh_RJ)
n_id <- sample(x = 3:5, size = n_reg, replace = TRUE)
coefs <- c(0.3, -0.3)
tau <- 0.75 # Scale of spatial effect

##-- Data
data <- rsurv(n_id = n_id,
              coefs = coefs, cens = 0.5, scale = FALSE,
              cens_type = "right", hazard = "weibull",
              hazard_params = list(weibull = list(alpha = 1.2, variant = 0)),
              spatial = "ICAR",
              neigh = neigh_RJ, tau = tau, confounding = "linear", proj = "none")

##-- Models
weibull_inla <- rsfm(data = data,
                     formula = surv(time = L, event = status) ~ X1 + X2,
                     family = "weibull", model = "none",
                     proj = "rhz", nsamp = 1000, approach = "inla")

rsfm_inla <- rsfm(data = data,
                  formula = surv(time = L, event = status) ~ X1 + X2,
                  family = "weibull", area = "reg",
                  model = "r_besag", neigh = neigh_RJ,
                  proj = "rhz", nsamp = 1000, approach = "inla")

weibull_inla$unrestricted$summary_fixed
rsfm_inla$unrestricted$summary_fixed
rsfm_inla$restricted$summary_fixed

}
