% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm_inla.R, R/rsglmm_inla.R
\name{rsfm_inla}
\alias{rsfm_inla}
\title{Restricted Spatial Frailty Model in INLA}
\usage{
rsfm_inla(f, data, family, W = NULL,
                 proj = "none", fast = TRUE, nsamp = 1000, ...)

rsfm_inla(
  f,
  data,
  family,
  W = NULL,
  proj = "none",
  fast = TRUE,
  nsamp = 1000,
  ...
)
}
\arguments{
\item{f}{INLA formula ?inla.surv}

\item{data}{data.frame containing, at least, \code{time}, \code{status}, \code{covariates}, \code{area} list}

\item{family}{'exponential', 'weibull', 'weibullcure', 'loglogistic', 'gamma', 'lognormal' or 'pwe'}

\item{W}{Adjacency matrix}

\item{proj}{'none', 'rhz' or 'spock'}

\item{fast}{Should we use the reduction operator?}

\item{nsamp}{Sample size to use the projection approach}

\item{...}{Other parameters used in ?inla}
}
\value{
INLA object with corrected parameters

INLA object with corrected parameters
}
\description{
Fit a Restricted Spatial Frailty model using INLA

Fit a Restricted Spatial Frailty model using INLA
}
