% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfm_mcmc.R
\name{rsfm_mcmc}
\alias{rsfm_mcmc}
\title{Restricted Spatial Frailty Model in BUGS}
\usage{
rsfm_mcmc(model, data, inits, parameters, covariates, area,
                 proj = "none", fast = TRUE, nsamp = 1000, burnin = 5000, lag = 10, ...)
}
\arguments{
\item{model}{character with the model in text format.}

\item{data}{an data frame or list containing the variables in the model.}

\item{inits}{initial values for the chain.}

\item{parameters}{vector of parameters names to restore.}

\item{covariates}{vector of covariates names.}

\item{area}{areal variable name in \code{data}.}

\item{proj}{'none' or 'rhz'.}

\item{fast}{TRUE to use the reduction operator.}

\item{nsamp}{number of desired. samples Default = 1000.}

\item{burnin}{burnin size.}

\item{lag}{lag parameter.}

\item{...}{other parameters used in ?R2OpenBUGS::bugs}
}
\value{
\item{$unrestricted}{A list containing
                               \itemize{
                                  \item $sample a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed summary measures for the coefficients
                                  \item $summary_hyperpar summary measures for hyperparameters
                                  \item $summary_random summary measures for random quantities
                                }
                             }
\item{$restricted}{A list containing
                               \itemize{
                                  \item $sample a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed summary measures for the coefficients
                                  \item $summary_hyperpar summary measures for hyperparameters
                                  \item $summary_random summary measures for random quantities
                                }
                             }

\item{$out}{BUGS output}
\item{$time}{time elapsed for fitting the model}
}
\description{
Fit a Restricted Spatial Frailty model using BUGS
}
