% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsglmm.R
\name{rsglmm}
\alias{rsglmm}
\title{Restricted Spatial Generalized Linear Mixed model}
\usage{
rsglmm(data, formula, family,
              E = NULL, n = NULL,
              area = NULL, model = NULL, neigh = NULL,
              proj = "none", nsamp = 1000, burnin = 5000, lag = 1,
              priors = list(prior_prec = list(tau = c(0.5, 0.0005))),
              approach = "inla",
              ...)
}
\arguments{
\item{data}{an data frame or list containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{some allowed families are: 'gaussian', 'poisson' and 'binomial'. The family availability will depend on the approach.}

\item{E}{known component, in the mean for the Poisson likelihoods defined as E = exp(\eqn{\eta}), where \eqn{\eta} is the linear predictor. Default = 1.}

\item{n}{a vector containing the number of trials for the binomial likelihood, or the number of required successes for the nbinomial2 likelihood. Default value is set to 1.}

\item{area}{areal variable name in \code{data}.}

\item{model}{spatial model adopted. Examples: "besag", "besag2" or "r_besag". See INLA::inla.list.models() for other models.}

\item{neigh}{neighborhood structure. A \code{SpatialPolygonsDataFrame} object.}

\item{proj}{"none", "rhz", "hh" or "spock"}

\item{nsamp}{number of samples. Default = 1000.}

\item{burnin}{burn-in size (just for hh).}

\item{lag}{lag parameter (just for hh).}

\item{priors}{a list containing (for now):
\itemize{
   \item prior_prec: a list with:
   \itemize{
       \item tau: a vector of size two containing shape and scale for the gamma distribution applied for \eqn{\tau}
   }
}}

\item{approach}{'inla' or 'mcmc'}

\item{...}{other parameters used in ?INLA::inla or ?ngspatial::sparse.sglmm}
}
\value{
\item{$unrestricted}{A list containing
                               \itemize{
                                  \item $sample: a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed: summary measures for the coefficients
                                  \item $summary_hyperpar: summary measures for hyperparameters
                                  \item $summary_random: summary measures for random quantities
                                }
                             }
\item{$restricted}{A list containing
                               \itemize{
                                  \item $sample: a sample of size nsamp for all parameters in the model
                                  \item $summary_fixed: summary measures for the coefficients
                                  \item $summary_hyperpar: summary measures for hyperparameters
                                  \item $summary_random: summary measures for random quantities
                                }
                             }

\item{$out}{INLA or ngspatial output}
\item{$time}{time elapsed for fitting the model}
}
\description{
Fit a Restricted Spatial Generalized Linear Mixed model
}
\details{
The fitted model is given by
\deqn{Y ~ Poisson(E\theta),}

\deqn{log(\theta) = X\beta + \psi,}

\deqn{\psi ~ ICAR(\tau).}
}
\examples{
set.seed(123456)

##-- Spatial structure
data("neigh_RJ")

beta <- c(-0.5, -0.2)
tau <- 1

##-- Data ----
family <- "poisson"
data <- rglmm(beta = beta, tau = tau, family = family,
              confounding = "none", neigh = neigh_RJ,
              scale = TRUE)

##-- Models ----
sglm_mod <- rsglmm(data = data, formula = Y ~ X1 + X2,
                   family = family,
                   proj = "none", nsamp = 1000)

sglmm_mod <- rsglmm(data = data, formula = Y ~ X1 + X2,
                    family = family,
                    area = "reg", model = "besag", neigh = neigh_RJ,
                    proj = "none", nsamp = 1000)

rglmm_rhz <- rsglmm(data = data, formula = Y ~ X1 + X2,
                    family = family,
                    area = "reg", model = "r_besag", neigh = neigh_RJ,
                    proj = "rhz", nsamp = 1000)

rglmm_spock <- rsglmm(data = data, formula = Y ~ X1 + X2,
                      family = family,
                      area = "reg", model = "r_besag", neigh = neigh_RJ,
                      proj = "spock", nsamp = 1000)

rglmm_hh <- rsglmm(data = data, formula = Y ~ X1 + X2,
                   family = family,
                   area = "reg", model = "r_besag", neigh = neigh_RJ,
                   approach = "mcmc",
                   proj = "hh", burnin = 5000, nsamp = 1000, lag = 10)

sglm_mod$unrestricted$summary_fixed
sglmm_mod$unrestricted$summary_fixed
rglmm_rhz$unrestricted$summary_fixed
rglmm_rhz$restricted$summary_fixed
rglmm_spock$restricted$summary_fixed
rglmm_hh$restricted$summary_fixed

}
