% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsglmm_inla.R
\name{rsglmm_inla}
\alias{rsglmm_inla}
\title{Restricted Spatial Generalized Linear Mixed model in INLA}
\usage{
rsglmm_inla(f, data, family, W = NULL, proj = "none", nsamp = 1000, ...)
}
\arguments{
\item{f}{INLA formula ?inla.surv}

\item{data}{data.frame containing, at least, \code{time}, \code{status}, \code{covariates}, \code{area} list}

\item{family}{'exponential', 'weibull', 'weibullcure', 'loglogistic', 'gamma', 'lognormal' or 'pwe'}

\item{W}{Adjacency matrix}

\item{proj}{'none', 'rhz' or 'spock'}

\item{nsamp}{Sample size to use the projection approach}

\item{...}{Other parameters used in ?inla}
}
\value{
INLA object with corrected parameters
}
\description{
Fit a Restricted Spatial Generalized Linear Mixed model using INLA
}
