% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsglmm_mcmc.R
\name{rsglmm_mcmc}
\alias{rsglmm_mcmc}
\title{Restricted Spatial Generalized Linear Mixed model}
\usage{
rsglmm_mcmc(
  data,
  formula,
  family,
  W,
  area,
  proj,
  nsamp = 1000,
  burnin = 5000,
  thin = 1,
  attractive = round(0.5 * (nrow(W)/2)),
  ...
)
}
\arguments{
\item{data}{data.frame containing, at least, \code{time}, \code{status}, \code{covariates}, \code{area} list}

\item{formula}{INLA formula ?inla.surv}

\item{family}{'exponential', 'weibull', 'weibullcure', 'loglogistic', 'gamma', 'lognormal' or 'pwe'}

\item{W}{Adjacency matrix}

\item{area}{Areal variable name in data}

\item{proj}{'none', 'rhz' or 'spock'}

\item{nsamp}{Sample size to use the projection approach}

\item{burnin}{Burnin period}

\item{thin}{Lag parameter}

\item{attractive}{The number of attractive Moran eigenvectors to use}

\item{...}{Other parameters used in ?ngspatial::sparse.sglmm}
}
\value{
INLA object with corrected parameters
}
\description{
Fit a Restricted Spatial Generalized Linear Mixed model using ngspatial
}
