% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sc.R
\name{rshared}
\alias{rshared}
\title{rshared}
\usage{
rshared(
  alpha_1 = 0,
  alpha_2 = 0,
  beta_1 = c(0.1, -0.1),
  beta_2 = c(0.1, -0.1),
  delta = 1,
  tau_1 = 1,
  tau_2 = 1,
  tau_s = 1,
  confounding = "none",
  neigh,
  scale = TRUE
)
}
\arguments{
\item{alpha_1}{intercept for disease 1.}

\item{alpha_2}{intercept for disease 2.}

\item{beta_1}{regression parameter(s) for disease 1.}

\item{beta_2}{regression parameter(s) for disease 2.}

\item{delta}{dependence on the shared component.}

\item{tau_1}{precision of ICAR specific to disease 1.}

\item{tau_2}{precision of ICAR specific to disease 2.}

\item{tau_s}{precision for ICAR specific to the shared component.}

\item{confounding}{"none", "linear", "quadratic" or "cubic".}

\item{neigh}{neighborhood structure. A \code{SpatialPolygonsDataFrame} object.}

\item{scale}{scale covariates? TRUE or FALSE.}
}
\description{
Generating data from Shared Component model
}
