% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sc.R
\name{rshared}
\alias{rshared}
\title{rshared}
\usage{
rshared(
  alpha_1 = 0,
  alpha_2 = 0,
  beta_1 = c(0.1, -0.1),
  beta_2 = c(0.1, -0.1),
  delta = 1,
  tau_1 = 1,
  tau_2 = 1,
  tau_s = 1,
  confounding = "none",
  neigh,
  scale = TRUE
)
}
\arguments{
\item{alpha_1}{Intercept for disease 1}

\item{alpha_2}{Intercept for disease 2}

\item{beta_1}{Coefficients for disease 1}

\item{beta_2}{Coefficients for disease 2}

\item{delta}{Dependence on the shared component}

\item{tau_1}{Precision for ICAR specific for disease 2}

\item{tau_s}{Precision for ICAR specific for shared component}

\item{confounding}{'none', 'linear', 'quadratic' or 'cubic'}

\item{neigh}{Neighborhood structure. A \code{SpatialPolygonsDataFrame} object}

\item{scale}{Scale covariates? TRUE or FALSE}
}
\description{
Generating data from Shared Component model
}
