% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rsurv}
\alias{rsurv}
\title{rsurv}
\usage{
rsurv(
  n_id,
  coefs = c(0.1, 0.4, -0.3),
  cens = 0,
  cens_type = "interval",
  hazard = "weibull",
  hazard_params = hazard_dft(),
  spatial = "ICAR",
  neigh = NULL,
  tau = 1,
  confounding = "none",
  proj = "none",
  sd_x = 0,
  X = NULL,
  scale = TRUE
)
}
\arguments{
\item{n_id}{Vector with the number of individuals in each region to be generated}

\item{coefs}{Vector of coefficients}

\item{cens}{Censure level}

\item{cens_type}{Censoring scheme: 'left', 'right' or ''interval}

\item{hazard}{Hazard model: 'exponencial', 'weibull' or 'pwe'}

\item{hazard_params}{Named list with parameters for the hazard model: hazard_dft()}

\item{spatial}{Spatial model: 'none' for the conventional Cox model,
'gamma' for an independent gamma frailty,
'lognormal' for an independent lognormal frailty,
'ICAR' or 'BYM' for spatial structured models}

\item{neigh}{Neighborhood structure. A \code{SpatialPolygonsDataFrame} object}

\item{tau}{Precision for ICAR and BYM models}

\item{confounding}{'none', 'linear', 'quadratic' or 'cubic'}

\item{proj}{'none', 'rhz', 'hh' or 'spock'}

\item{sd_x}{Standard deviation to generating confounding}

\item{X}{Matrix of covariates. Default = NULL}

\item{scale}{Scale X. TRUE or FALSE}
}
\description{
Generating data with spatial confounding
}
