% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rsurv_left}
\alias{rsurv_left}
\title{rsurv_left}
\usage{
rsurv_left(times, cens)
}
\arguments{
\item{times}{vector with observed times.}

\item{cens}{censoring fraction.}
}
\description{
Generating data with left censure
}
