% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rsurv_right}
\alias{rsurv_right}
\title{rsurv_right}
\usage{
rsurv_right(times, cens)
}
\arguments{
\item{times}{observed time vector.}

\item{cens}{censored proportion.}
}
\description{
Generating data with right censoring
}
\keyword{internal}
