% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{rweibullsurv}
\alias{rweibullsurv}
\title{rweibullsurv}
\usage{
rweibullsurv(N, alpha, lambda = rep(1, N), variant = 1)
}
\arguments{
\item{N}{number of observations.}

\item{alpha}{shape parameter.}

\item{lambda}{scale parameter.}

\item{variant}{variant (0 or 1).}
}
\description{
Generate data from Weibull hazard model
}
\keyword{internal}
