% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slovenia.R
\docType{data}
\name{slovenia}
\alias{slovenia}
\title{Slovenia stomach cancer dataset}
\format{An object of class \code{"sf"}.}
\usage{
data(slovenia)
}
\description{
Slovenia stomach cancer dataset
}
\references{
Zadnik, Vesna, and B. J. Reich. \emph{"Analysis of the relationship between socioeconomic factors and stomach cancer incidence in Slovenia"}. Neoplasma 53.2 (2006): 103-110.
}
\keyword{datasets}
