% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spock.R
\name{spock}
\alias{spock}
\title{SPOCK nb}
\usage{
spock(X, map)
}
\arguments{
\item{X}{covariate matrix}

\item{map}{neighborhood object}
}
\value{
map_new new neighborhood object
}
\description{
Apply the SPOCK approach to transform the neighborhood matrix
}
