% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surv.R
\name{surv}
\alias{surv}
\title{surv}
\usage{
surv(time, time2 = NULL, event)
}
\arguments{
\item{time}{time to event (right censoring), or lower limit (interval censoring).}

\item{time2}{upper limit (interval censoring).}

\item{event}{event indicator; 1 = observed event, 0 = right censored event, 2 = left censored event, 3 = interval censored event.}
}
\description{
Auxiliar function for survival models
}
\keyword{internal}
