% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVIF}
\alias{SVIF}
\title{Spatial Variance Inflation Factor}
\usage{
SVIF(base_model, model)
}
\arguments{
\item{base_model}{model to use as comparison.}

\item{model}{model to compare.}
}
\value{
res data.frame with SVIF for fixed parameters
}
\description{
Calculate the spatial variance inflation factor (SVIF)
}
