% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVRF}
\alias{SVRF}
\title{Spatial Variance Retraction Factor}
\usage{
SVRF(base_model, model)
}
\arguments{
\item{base_model}{model to use as comparison}

\item{model}{model to compare}
}
\value{
res data.frame with SVRF for fixed parameters
}
\description{
Calculate the spatial variance retraction factor (SVRF) that lies between -1 (inflaction) and 1 (rettraction)
}
