% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WAIC}
\alias{WAIC}
\title{Watanabe–Akaike information criterion}
\usage{
WAIC(object)
}
\arguments{
\item{object}{a object from ?rsglmm, ?rscm or ?rsfm}
}
\value{
WAIC
}
\description{
Get the Watanabe–Akaike information criterion (WAIC) from a model
}
\references{
Watanabe, S. (2010). \emph{"Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory"}. Journal of Machine Learning Research, 11(Dec), 3571-3594.
}
